Sponge update and some bytecodeclass stuff

+++ src/main/java/org/granitepowered/granite/impl/entity/GraniteEntityExperienceOrb.java
@@ -33,12 +33,12 @@ public class GraniteEntityExperienceOrb extends GraniteEntity<MCEntityXPOrb> imp
     }
 
     @Override
-    public double getExperience() {
+    public int getExperience() {
         return obj.fieldGet$xpValue();
     }
 
     @Override
-    public void setExperience(double experience) {
-        obj.fieldSet$xpValue((int) experience);
+    public void setExperience(int experience) {
+        obj.fieldSet$xpValue(experience);
     }
 }
